+++ frontend/src/pages/CartsZones.tsx
@@ -1,9 +1,0 @@
-import ZonesTab from "./CartsComponents/ZonesTab";
-
-export default function CartsZones() {
-  return (
-    <div className="animate-in fade-in duration-300">
-      <ZonesTab />
-    </div>
-  );
-}